API_FastAPI/NewsApp/requirements.txt
flask
newsapi-python
